requirements.txt
selenium
openai
requests
python-dotenv
pdfminer.six